fix(slm): wire fleet sync concurrent guard into schedule paths (#1979)
Extract assert_no_running_sync + fleet_sync_lock into shared
services/fleet_sync_guard.py to avoid circular imports. Wire guard
into run_schedule endpoint and execute_schedule background executor.
Removes duplicate inline guard that lacked lock protection.
